refactor: 인수 테스트 클래스 개선 - 인수 테스트 공통 역할 클래스로 추출하여 각 인수테스트에서 확장하여 사용하도록 분리 - LineAcceptanceTest 내 라인 생성 request 공통 객체 필드로 선언하여 추출 - 라인 생성 요청시 request body 를 map 대신 서비스에서 사용하는 LineRequest DTO 사용하여 생성하도록 테스트 변경
