+++ src/main/java/nextstep/subway/dto/LineRequest.java
@@ -10,6 +10,15 @@ public class LineRequest {
     private long downStationId;
     private long distance;
 
+    public LineRequest() {}
+    public LineRequest(String name, String color, long upStationId, long downStationId, long distance) {
+        this.name = name;
+        this.color = color;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
     public Line toLine(Station upStation, Station downStation) {
         return new Line(name, color, upStation, downStation, distance);
     }
